Add support for using open structures as targets.
This involves a number of changes in the algorithm at places where the orientation 
of a structure or checks whether points are inside a structure are used. When closed
structures are used as targets, the algorithm falls back on the traditional approach.

This caused some refactoring in the structure handling routines.

An important change is that CarreStructures%npstru is now positive for all structures, also open ones.
An additional logical flag CarreStructures%closed stores whether the structure was given as closed
in the input file.



git-svn-id: http://solps-mdsplus.aug.ipp.mpg.de/repos/CARRE/branches/hmk@238 4d76dbe9-ff4f-0410-8810-ffbc1acdba1b